+++ script/coverage.php
@@ -15,7 +15,7 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 use cloak\Analyzer;
 use cloak\ConfigurationBuilder;
-use cloak\driver\result\File;
+use cloak\driver\result\FileResult;
 use cloak\reporter\CompositeReporter;
 use cloak\reporter\ProcessingTimeReporter;
 use cloak\reporter\TextReporter;
@@ -34,9 +34,9 @@ $analyzer = Analyzer::factory(function(ConfigurationBuilder $builder) {
         new TextReporter(),
         new ProcessingTimeReporter()
     ]));
-    $builder->includeFile(function(File $file) {
+    $builder->includeFile(function(FileResult $file) {
         return $file->matchPath('/src');
-    })->excludeFile(function(File $file) {
+    })->excludeFile(function(FileResult $file) {
         return $file->matchPath('/spec') || $file->matchPath('/vendor');
     });