+++ spec/reflection/ClassReflectionSpec.php
@@ -10,8 +10,9 @@
  */
 
 use cloak\reflection\ClassReflection;
-use cloak\result\LineSet;
-use cloak\result\Line;
+use cloak\result\collection\LineResultCollection;
+use cloak\result\LineResult;
+
 
 describe('ClassReflection', function() {
     before(function() {
@@ -66,8 +67,8 @@ describe('ClassReflection', function() {
     describe('assembleBy', function() {
         context('when class reflection', function() {
             before(function() {
-                $result = $this->classReflection->assembleBy(new LineSet([
-                    new Line(29, Line::UNUSED)
+                $result = $this->classReflection->assembleBy(new LineResultCollection([
+                    new LineResult(29, LineResult::UNUSED)
                 ]));
                 $this->result = $result;
             });
@@ -85,8 +86,8 @@ describe('ClassReflection', function() {
         });
         context('when trait reflection', function() {
             before(function() {
-                $result = $this->traitReflection->assembleBy(new LineSet([
-                    new Line(11, Line::UNUSED)
+                $result = $this->traitReflection->assembleBy(new LineResultCollection([
+                    new LineResult(11, LineResult::UNUSED)
                 ]));
                 $this->result = $result;
             });