+++ spec/result/collection/NamedResultCollectionSpec.php
@@ -11,17 +11,17 @@
 
 use cloak\result\collection\NamedResultCollection;
 use cloak\result\type\ClassResult;
-use cloak\result\LineSet;
-use cloak\result\Line;
+use cloak\result\collection\LineResultCollection;
+use cloak\result\LineResult;
 use cloak\reflection\ClassReflection;
 
 
 describe('ClassResultCollection', function() {
     describe('#add', function() {
         before(function() {
-            $lineSet = new LineSet([
-                new Line(12, Line::EXECUTED),
-                new Line(17, Line::UNUSED)
+            $lineSet = new LineResultCollection([
+                new LineResult(12, LineResult::EXECUTED),
+                new LineResult(17, LineResult::UNUSED)
             ]);
             $classReflection = new ClassReflection('Example\\Example');
 
@@ -34,9 +34,9 @@ describe('ClassResultCollection', function() {
     });
     describe('#getIterator', function() {
         before(function() {
-            $lineSet = new LineSet([
-                new Line(12, Line::EXECUTED),
-                new Line(17, Line::UNUSED)
+            $lineSet = new LineResultCollection([
+                new LineResult(12, LineResult::EXECUTED),
+                new LineResult(17, LineResult::UNUSED)
             ]);
             $classReflection = new ClassReflection('Example\\Example');