+++ src/reflection/FileReflection.php
@@ -13,8 +13,8 @@ namespace cloak\reflection;
 
 use cloak\reflection\collection\ReflectionCollection;
 use cloak\value\LineRange;
-use cloak\result\File;
-use cloak\result\LineSetInterface;
+use cloak\result\FileResult;
+use cloak\result\LineResultCollectionInterface;
 use PhpCollection\Sequence;
 use Zend\Code\Reflection\ClassReflection as ZendClassReflection;
 use Zend\Code\Reflection\FileReflection as ZendFileReflection;
@@ -106,12 +106,12 @@ class FileReflection implements ReflectionInterface
     }
 
     /**
-     * @param LineSetInterface $lineResults
-     * @return File
+     * @param LineResultCollectionInterface $lineResults
+     * @return FileResult
      */
-    public function assembleBy(LineSetInterface $lineResults)
+    public function assembleBy(LineResultCollectionInterface $lineResults)
     {
-        return new File($this->getName(), $lineResults);
+        return new FileResult($this->getName(), $lineResults);
     }
 
 }